[FIX]: handle SIGPIPE (broken pipe)
종료된 자식프로세스 pipe write 시도 시 SIGPIPE 발생하여
signal() 및 프로세스 상태 체크 추가

+++ includes/Type.hpp
@@ -86,7 +86,7 @@ class PipeFD : public FDType
 {
 	private:
 		const std::string &data;
-		int write_idx;
+		size_t write_idx;
 		pid_t pid;
 		Client *client;
 
@@ -97,9 +97,9 @@ class PipeFD : public FDType
 		Client *getClient(void);
 		pid_t getPid(void);
 		const std::string &getData();
-		int getWriteIdx();
+		size_t getWriteIdx();
 
-		void setWriteIdx(int write_idx);
+		void setWriteIdx(size_t write_idx);
 };
 
 void setFDonTable(int fd, t_fdset set, FDType *data);
